requirements.txt
mkdocs-material
mkdocs-awesome-nav